+++ app/actions/car-onboarding/registration-certificate-not-recognized.error.ts
@@ -1,6 +1,0 @@
-export class RegistrationCertificateNotRecognizedError extends Error {
-  constructor() {
-    super('The uploaded image could not be recognized as a registration certificate front side');
-    this.name = 'RegistrationCertificateNotRecognizedError';
-  }
-}